backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.128.0"}
psycopg = {extras = ["binary"], version = "^3.3.2"}
pyjwt = "^2.10.1"
python-multipart = "^0.0.21"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiosqlite = "^0.22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
